+++ internal/geoipupdate/config.go
@@ -364,7 +364,10 @@ func setConfigFromEnv(config *Config) error {
 	}
 
 	if value, ok := os.LookupEnv("GEOIPUPDATE_VERBOSE"); ok {
-		config.Verbose = value != ""
+		if value != "0" && value != "1" {
+			return errors.New("`GEOIPUPDATE_VERBOSE' must be 0 or 1")
+		}
+		config.Verbose = value == "1"
 	}
 
 	return nil